feat(api): centralizar validacao de payload em sources
