+++ src/handlers/update-source.ts
@@ -1,44 +1,18 @@
 import {
   SourceVersionConflictError,
-  type SourceRegistryRecord,
   type SourceRegistryRepository,
 } from '../domain/sources/source-registry-repository';
 import {
-  type SourceSchemaValidationError,
-  validateSourceSchemaV1,
-} from '../domain/sources/source-schema';
+  mergeAndValidateSourcePatch,
+  SOURCE_PAYLOAD_VALIDATION_MESSAGE,
+  validateSourcePatchPayload,
+} from '../domain/sources/source-payload-validation';
 import { createDynamoDbSourceRegistryRepository } from '../infra/sources/dynamodb-source-registry-repository';
 import { nowIso } from '../shared/time/now-iso';
 
 const JSON_HEADERS = {
   'content-type': 'application/json',
 };
-
-const IMMUTABLE_FIELDS = ['sourceId', 'engine', 'schemaVersion', 'createdAt', 'updatedAt'] as const;
-const MUTABLE_FIELDS = [
-  'active',
-  'secretArn',
-  'query',
-  'cursorField',
-  'fieldMap',
-  'scheduleType',
-  'intervalMinutes',
-  'cronExpr',
-  'nextRunAt',
-] as const;
-const KNOWN_FIELDS = new Set<string>([...IMMUTABLE_FIELDS, ...MUTABLE_FIELDS]);
-
-interface UpdateSourcePatchPayload {
-  active?: boolean;
-  secretArn?: string;
-  query?: string;
-  cursorField?: string;
-  fieldMap?: Record<string, string>;
-  scheduleType?: 'interval' | 'cron';
-  intervalMinutes?: number;
-  cronExpr?: string;
-  nextRunAt?: string;
-}
 
 export interface UpdateSourceEvent {
   body?: string | null;
@@ -68,9 +42,6 @@ const response = (statusCode: number, payload: unknown): UpdateSourceResponse =>
   headers: JSON_HEADERS,
   body: JSON.stringify(payload),
 });
-
-const isRecord = (value: unknown): value is Record<string, unknown> =>
-  typeof value === 'object' && value !== null && !Array.isArray(value);
 
 const parseBody = (
   rawBody: string | null | undefined,
@@ -117,116 +88,6 @@ const parseSourceId = (
   };
 };
 
-const validatePatchPayload = (
-  payload: unknown,
-):
-  | { success: true; value: UpdateSourcePatchPayload }
-  | { success: false; errors: SourceSchemaValidationError[] } => {
-  if (!isRecord(payload)) {
-    return {
-      success: false,
-      errors: [
-        {
-          field: '$',
-          code: 'INVALID_TYPE',
-          message: 'patch payload must be an object.',
-        },
-      ],
-    };
-  }
-
-  const errors: SourceSchemaValidationError[] = [];
-  const patch: UpdateSourcePatchPayload = {};
-  let mutableFieldCount = 0;
-
-  for (const key of Object.keys(payload)) {
-    if (!KNOWN_FIELDS.has(key)) {
-      errors.push({
-        field: key,
-        code: 'INVALID_VALUE',
-        message: `field "${key}" is not supported for source updates.`,
-      });
-      continue;
-    }
-
-    if ((IMMUTABLE_FIELDS as readonly string[]).includes(key)) {
-      errors.push({
-        field: key,
-        code: 'INVALID_VALUE',
-        message: `field "${key}" is immutable and cannot be updated.`,
-      });
-      continue;
-    }
-
-    mutableFieldCount += 1;
-    patch[key as keyof UpdateSourcePatchPayload] = payload[key] as never;
-  }
-
-  if (mutableFieldCount === 0) {
-    errors.push({
-      field: '$',
-      code: 'INVALID_VALUE',
-      message: 'patch payload must include at least one mutable field.',
-    });
-  }
-
-  if (errors.length > 0) {
-    return {
-      success: false,
-      errors,
-    };
-  }
-
-  return {
-    success: true,
-    value: patch,
-  };
-};
-
-const mergeSourceRecord = (
-  current: SourceRegistryRecord,
-  patch: UpdateSourcePatchPayload,
-  nextUpdatedAt: string,
-): SourceRegistryRecord | { validationErrors: SourceSchemaValidationError[] } => {
-  const nextScheduleType = patch.scheduleType ?? current.scheduleType;
-  const nextIntervalMinutes =
-    nextScheduleType === 'interval'
-      ? (patch.intervalMinutes ??
-        (current.scheduleType === 'interval' ? current.intervalMinutes : undefined))
-      : patch.intervalMinutes;
-  const nextCronExpr =
-    nextScheduleType === 'cron'
-      ? (patch.cronExpr ?? (current.scheduleType === 'cron' ? current.cronExpr : undefined))
-      : patch.cronExpr;
-
-  const validation = validateSourceSchemaV1({
-    sourceId: current.sourceId,
-    active: patch.active ?? current.active,
-    engine: current.engine,
-    secretArn: patch.secretArn ?? current.secretArn,
-    query: patch.query ?? current.query,
-    cursorField: patch.cursorField ?? current.cursorField,
-    fieldMap: patch.fieldMap ?? current.fieldMap,
-    scheduleType: nextScheduleType,
-    intervalMinutes: nextIntervalMinutes,
-    cronExpr: nextCronExpr,
-    nextRunAt: patch.nextRunAt ?? current.nextRunAt,
-  });
-
-  if (!validation.success) {
-    return {
-      validationErrors: validation.errors,
-    };
-  }
-
-  return {
-    ...validation.value,
-    schemaVersion: current.schemaVersion,
-    createdAt: current.createdAt,
-    updatedAt: nextUpdatedAt,
-  };
-};
-
 const getDefaultDependencies = (): UpdateSourceDependencies => {
   if (cachedDefaultDependencies) {
     return cachedDefaultDependencies;
@@ -258,10 +119,10 @@ export const createHandler =
       return parsedBody.response;
     }
 
-    const patchValidation = validatePatchPayload(parsedBody.value);
+    const patchValidation = validateSourcePatchPayload(parsedBody.value);
     if (!patchValidation.success) {
-      return response(422, {
-        message: 'Source patch validation failed.',
+      return response(400, {
+        message: SOURCE_PAYLOAD_VALIDATION_MESSAGE,
         errors: patchValidation.errors,
       });
     }
@@ -275,27 +136,27 @@ export const createHandler =
     }
 
     const nextUpdatedAt = now();
-    const merged = mergeSourceRecord(current, patchValidation.value, nextUpdatedAt);
-    if ('validationErrors' in merged) {
-      return response(422, {
-        message: 'Source patch validation failed.',
-        errors: merged.validationErrors,
+    const merged = mergeAndValidateSourcePatch(current, patchValidation.value, nextUpdatedAt);
+    if (!merged.success) {
+      return response(400, {
+        message: SOURCE_PAYLOAD_VALIDATION_MESSAGE,
+        errors: merged.errors,
       });
     }
 
     try {
       await sourceRegistryRepository.update({
         sourceId: current.sourceId,
-        source: merged,
+        source: merged.value,
         expectedUpdatedAt: current.updatedAt,
       });
 
       return response(200, {
-        sourceId: merged.sourceId,
+        sourceId: merged.value.sourceId,
         metadata: {
-          schemaVersion: merged.schemaVersion,
-          createdAt: merged.createdAt,
-          updatedAt: merged.updatedAt,
+          schemaVersion: merged.value.schemaVersion,
+          createdAt: merged.value.createdAt,
+          updatedAt: merged.value.updatedAt,
           requestId: event.requestContext?.requestId ?? null,
         },
       });
